+++ web/src/app/eros.proto.js
@@ -490,7 +490,7 @@
                     "id": 3
                 },
                 {
-                    "rule": "required",
+                    "rule": "optional",
                     "options": {},
                     "type": "int64",
                     "name": "opponent_latency",