+++ web/src/app/liberos/modules/enums.js
@@ -72,6 +72,23 @@
     		NOSHOW: "AQ==",
     		DRAW: "Ag=="
     	}
+
+    	this.ErrorDescriptor = {
+		    301: "Error processing replay",
+		    302: "Error while processing match result",
+		    303: "Duplicate Replay",
+		    304: "The submitting client was not involved in the match.",
+		    305: "Game too short.",
+		    306: "Bad format. Required 1v1 with no observers.",
+		    307: "Bad map. Require a map in the map pool.",
+		    308: "All participants of the game must be registered.",
+		    309: "Player not found in database.",
+		    310: "You didn't play your matchmade opponent. You have been forfeited from that game.",
+		    311: "The game was not played on Faster.",
+		    312: "Cannot add veto. Map not in ranked pool.",
+		    313: "Cannot add veto. Maximum number of vetoes used.",
+		    314: "You are not in a game arranged by the Eros matchmaker."
+    	}
     }
 
     global.starbow.Eros.prototype.modules.enums = EnumsModule;