+++ web/src/app/modules/eros-chat/directives.js
@@ -105,6 +105,13 @@ angular.module('erosApp.chat', ['ngAudio'])
 		restrict: 'A',
 		link: function($scope, $elem, $attrs, $controller){
 			$elem.html($scope.message.message.replace(/@\w+/g, '<span username>'+'$&'+'</span>'));
+			$elem.html($scope.message.message.replace(/http\S*/g, function(match){
+				var rep = '<a href="'+match+'" target="_blank">';
+				rep += match.length > 40 ? match.slice(0,30)+'...'+match.slice(-6) : match;
+				rep += '</a>'
+
+				return rep;
+			}));
 		}
 	};
 })
@@ -153,7 +160,7 @@ angular.module('erosApp.chat', ['ngAudio'])
 	return function(division){
 		switch(division) {
 		case 'E':
-			return "#888";
+			return "#9966cc";
 			break
 		case 'D':
 			return "#AB3E3E";